refactor(tool_input_repair): extract hints/semantic/validate/error_fmt (dirge-4y4l stage 10b)
Final structural stage for the tool_input_repair god-file. Splits the
remaining four concern clusters out of mod.rs into sibling modules, each
re-exported via `pub use <mod>::*` so external paths
(crate::agent::agent_loop::tool_input_repair::*) stay unchanged.

- hints.rs (227) — contract_hint_for + with_contract_hint + the inline
  phase2_tests module (its `use super::*` -> `use super::super::*` to keep
  reaching the crate-level re-exports).
- semantic.rs (244) — schema-driven semantic repair: apply_relational_defaults,
  MD_LINK_RE/unwrap_md_link, SemanticTag/extract_semantic_tag, is_path_field,
  unwrap_md_links_in_args.
- validate.rs (266) — validate_and_repair orchestration + strip_null_*,
  apply_repair_at_*, try_repairs_at_value, parse_json_pointer.
- error_fmt.rs (103) — format_structured_error, extract_schema_hint,
  navigate_schema, build_concrete_hint, is_path_field_name.
- mod.rs (190, was 985) — telemetry types (RepairKind/RepairResult/
  RepairStats/RepairStatsSnapshot) + module wiring only.

Cross-module helpers made pub(super) (semantic::{apply_relational_defaults,
unwrap_md_links_in_args,unwrap_md_link}, validate::parse_json_pointer,
error_fmt::{navigate_schema,build_concrete_hint}); the dependency cycle
semantic->error_fmt->validate->semantic is fine across modules. tests.rs
imports the pub(super) internals it exercises directly.

End state: tool_input_repair.rs 2035 -> 7 files, largest non-test 266 LOC.
-D warnings clean default + windows-default; 2374 tests green.

Refs dirge-u15f
